Add "word_suggestions" field to headword-and-forms search.
Previously, the headword-and-forms suggester set was
changed to use specially-constructed "word_suggestions"
based on all possible combinations of words constructed
from the headwod, e.g. "whit(e)" becomes "whit, white".

The things actually being searched, however, was NOT
including this new field. The result was that things
appearing in the dropdown were *not* actually showing
up when you clicked on an item to conduct the search.

This adds field "word_suggestions"
to the headword-and-forms search.