Build 2 containers and deploy automatically to testing
This modifies Dockerfile.prod so that it has two phases: webapp and
assets. The single phase that used to be there has been named webapp,
and the new one is a simple nginx container named assets.

(You'll have to add `-t webapp` or `-t assets` to `docker build` from
now on, when building by hand from this containerfile.)

We also add a github action workflow for building and pushing the new
images as well as deploying them in the cluster in the testing
namespace.

In order to allow github to build the two gems that were served by
gems.www.lib.umich.edu, rather than figure out a way to let github
authenticate, we just switched the gemfile to reference the git
repositories instead of gems.www.lib.umich.edu. In the future, we might
prefer to inline them.