in the process of replacing AnnoyingUtilities with SolrHelper
I think something is going wrong with data prepare currently. There's this
spot in the code in both prepare_new_data.rb and extract_convert_index.rb
where SolrHelper.data_dir is reassigned (copying what AnnoyingUtilities was
doing). Something seems to be going wrong with the data prep process as indicated
by a flood of "thing not found in hypbibid file" prior to indexing so that
looks like the next step to me. I don't understand WHY data_dir is being reassigned.
I was in the process of tracing everywhere it's called (in a clean copy of the repo)
to see where the new value becomes important.