Set up production build with arg for relative root
By building the production target, you will get an image containing the
assets, ready to run at a relative root of `/`. If you pass
RAILS_RELATIVE_ROOT_URL as a build arg, it will be burned into the image
environment and the assets.

When this matches the deployed environment, URLs will be generated at
runtime including the prefix, to the CDN host. They will then read
through to the live instance and be cached. Then, URLs embedded in CSS
will be relative, including the prefix, so they will stay on the CDN
domain.