Fix suggestion path with relative root
This is an extremely obscure bug (??) and workaround with engines and
relative root URLs, which has been a perennially difficult combination.

Basically, I can't find a situation where the existing behavior would
actually be correct, but everything related to relative roots has been
changed and "fixed" and "broken" many times over the past decade.

This works around the problem by ensuring there is a trailing slash on
the SCRIPT_NAME when the paths are computed for engine routes. Why it
seems to only happen for suggest_index_path in the view partials is
beyond me for now.