Add minimal Shrine + Uppy updater
This fits together all of the little pieces to use multi-part uploads
directly to S3, relieving the form rendering and temp file handling:

- Add Shrine
- Add AWS S3 SDK
- Add content disposition detection
- Add Uppy S3 multi part
- Add layout to load Uppy CSS from CDN in head section
- Load incoming bucket configuration from ENV
- Set up optional Compose env_file (.app.env; requires 2.24+)
- Condition the initialization on ENV vars being present
- Add routes for updates controller
- Add routes for presigned and multipart direct uploads
- Condition the routes on S3 ENV vars being present
- Add basic view to configure and show Uppy interface
- Load Uppy mjs from CDN inline in the view

It's not perfect, but it does work, and it does support pause/retry for
partially completed uploads. The file lands in the incoming directory
when the upload is fully complete.