Build suggesters "manually" against each solr replica
We were dealing with some instabilities in the k8s cluster in 2024.09 when we wanted
to release this, resulting nodes loosing track of zookeeper and suggesters not getting
built (calls would time out). Attempt to brute-force this by hitting the individual
solr URLs directly, taking advantage of knowledge about how many replicas and what
they're called that we'd really rather not have to know.

NOTE: Once it's determined we don't need this hack, tear the relevant code in
MedInstaller:IndexingSteps#index (after the second call to @build_collection.commit),
leaving only a single `rebuild_suggesters` and call it a day.

To run in once for each configured solr replica, we need the following:

  * ENV[DIRECT_URLS_TO_SOLR_REPLICAS]: A space-delimited set of urls of the form
    "http://solr-solrcloud-1:8083" or whatever. This is the same format as the
    generic (cluster-level) connection string found in ENV[SOLR_URL] (e.g., no trailing "solr")
  * A non-falsey value for ENV[MANUALLY_BUILD_SUGGESTERS] to enable it.