Add a middleware to force the HTTP Host
There appears to be a subtle interaction or bug between ingress-nginx
and Rack in this case. Even with the upstream-vhost set and
use-forwarded-headers unset, the Ingress host is being supplied as
X-Forwarded-Host. Even with use-forwarded-headers set, the
X-Forwarded-Host is getting the ingress host.

In Rack, the forwarded_authority takes precedence over the
host_authority and server_authority... So, here we try to ensure that
our value, supplied via process environment, is written into the
forwarded host environment variable at the first position in the
middleware stack.