Only check for solr alias changes on a timer
The implementation of `solr_cloud-connection` is purposefully chatty, not caching anything because during the admin cycle, you want to see your changes reflected immediatley and the number of operations is generally small.

My use of uncached values (checking the name of the collection underlying an alias) on what turned out to be basically every call was disastrous.

This PR creates an instance of [Concurrent::TimerTask](https://ruby-concurrency.github.io/concurrent-ruby/master/Concurrent/TimerTask.html) that runs the check,
 and updates if needed, every 20 seconds.

 Notable changes are:
   * All the logic about updates is moved into `config/initializers/solr_admin_cache.rb`
   * The formerly-recursive method of determining the underlying collection
     name (because it's legal, in general, to have aliass that point to other
     aliases) has been removed, since we just don't need it.
   * The cached values are stored in `Concurrent::Atom` instances in
     the Services module with everything else
   * `load_local_config.rb` basically just calls `#value` on the
     Services values.
   * The footer now shows, ridiculously, the time down to the second.
     This is purely to make testing easier, because seeing a change in
     the footer date is an easy way to know the change has been made.