Use a memoizing ttl library to reduce solr calls.
The implementation of `solr_cloud-connection` is purposefully chatty,
not caching anything because during the admin cycle, you want to see
your changes reflected immediatley and the number of operations is generally small.

My use of uncached values (checking the name of the collection underlying an alias) on what
turned out to be basically every call was disastrous.

This PR creates use `ttl_memoizeable` to memoize the two calls for data that can
change after a new upload: the hyp_to_bib data, and the date of indexing (for the
footer).

In theory we should just have some sort of a listener that tells us when
the underlying collection has changed and have that trigger things,
but we don't, so here we are.

 Notable changes are:
   * Two methods in load_local_config now memoize values for the
     hyperbib mapping and the name of the current underlying
     connection.
   * The formerly-recursive method of determining the underlying collection
     name (because it's legal, in general, to have aliass that point to other
     aliases) has been removed, since we just don't need it.
   * The footer now shows, ridiculously, the time down to the second.
     This is purely to make testing easier, because seeing a change in
     the footer date is an easy way to know the change has been made.